\name{pb.logit.test}
\alias{pb.logit.test}
\title{This is only for testing other variants of the logit filter.}
\usage{
pb.logit.test(y, X, m0 = rep(0, ncol(X)), C0 = diag(ncol(X)),
  G = diag(nrow(C0)), W = diag(nrow(C0)), numParticles = 1000,
  seed = NULL, version = 0)
}
\description{
This is only for testing other variants of the logit
filter.
}

